% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat.R
\name{mcmc_rhat}
\alias{mcmc_rhat}
\title{Plot R-hat values}
\usage{
mcmc_rhat(object, ..., size = NULL)
}
\arguments{
\item{object}{An object (either a vector of r-hat values
or a "brmsfit" object"}

\item{...}{Currently ignored.}

\item{size}{An optional value to override \code{\link[ggplot2]{geom_point}}'s
default size (for \code{mcmc_rhat}, \code{mcmc_neff}) or
\code{\link[ggplot2]{geom_line}}'s default size (for \code{mcmc_acf}).}
}
\description{
\code{mcmc_rhat()} is a generic that plots R-hat values.
}
\examples{
1 + 1
}
\seealso{
\code{\link[bayesplot:mcmc_rhat]{bayesplot::mcmc_rhat()}}
}
