% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SomeDataDoc.R
\docType{data}
\name{SomeData}
\alias{SomeData}
\title{Some Data}
\format{A data frame with 10 rows and 2 variables
\itemize{
\item \code{a} Integers from 1 to 10
\item \code{b} Normal random noise (form Norm(0,1))
}}
\usage{
data(SomeData)
}
\description{
This is just some example data.
}
\examples{
head(SomeData)
}
\keyword{datasets}
